Implemented all required functionality.
Navigating to /api/whoami results in the app responding with a JSON
object containing the user's IP address and accept-language and
user-agent strings.

The style of the comments originally in the file was also changed.

The .gitignore file was updated to include the .DS_Store files prevalent
on Mac machines.